# packages/backend/src/modules/auth/auth.controller.ts
import crypto from 'node:crypto';
import { SESSION_COOKIE_MAX_AGE, SESSION_COOKIE_NAME } from '@/common/constants';
import { TranslatableError } from '@/common/error/translatable-error';
import { CacheService } from '@/core/cache/cache.service';
import { ConfigurationService } from '@/core/config/configuration.service';
import { LoggerService } from '@/core/logger/logger.service';
import axios from 'axios';
import {
  BadRequestException,
  Body,
  Controller,
  Delete,
  ForbiddenException,
  Get,
  Patch,
  Post,
  Query,
  Req,
  Res,
  ServiceUnavailableException,
  UnauthorizedException,
  UseGuards,
} from '@nestjs/common';
import type { Request, Response } from 'express';
import { AuthGuard } from './auth.guard';
import { AuthService } from './auth.service';
import { UserRepository } from '@/modules/user/user.repository';
import { SessionManager } from './session.manager';
import {
  ChangePasswordBody,
  ChangeUsernameBody,
  CheckResetPasswordRequestDto,
  DisableTotpBody,
  GetTotpUriBody,
  GetTotpUriDto,
  LoginBody,
  LoginDto,
  PasswordResetCompleteBody,
  PasswordResetCompleteDto,
  PasswordResetRequestBody,
  PasswordResetRequestDto,
  PasswordResetVerifyResponseDto,
  PortalDesktopExchangeDto,
  RegisterBody,
  RegisterDto,
  ResetPasswordBody,
  ResetPasswordDto,
  SetupTotpBody,
  VerifyTotpBody,
} from './dto/auth.dto';
import { ApiResponse } from '@nestjs/swagger';
import {
  buildPortalDesktopDeepLink,
  type PortalDesktopExchange,
  type PortalSsoState,
  resolveSameOriginRedirectUrl,
  toDesktopRedirectPath,
} from './portal-sso';

@Controller('auth')
export class AuthController {
  constructor(
    private readonly authService: AuthService,
    private readonly logger: LoggerService,
    private readonly config: ConfigurationService,
    private readonly cache: CacheService,
    private readonly userRepository: UserRepository,
    private readonly sessionManager: SessionManager,
  ) {}

  private async setSessionCookie(res: Response, sessionId: string, req: Request) {
    const host = req.headers['x-forwarded-host'] as string | undefined;
    const proto = req.headers['x-forwarded-proto'] as string | undefined;
    const domain = this.authService.getCookieDomain(host);
    const secure = Boolean(domain) && proto === 'https';

    this.logger.debug('Request headers', req.headers);
    this.logger.debug('Setting session cookie', { host, domain, proto, secure });

    if (this.config.get('userSettings').experimental.insecureCookie) {
      this.logger.warn('WARNING: Using insecure cookies. This is not recommended for production environments.');
      res.cookie(SESSION_COOKIE_NAME, sessionId, { httpOnly: true, secure: false, sameSite: 'lax', maxAge: SESSION_COOKIE_MAX_AGE });
    } else {
      res.cookie(SESSION_COOKIE_NAME, sessionId, {
        httpOnly: true,
        secure,
        sameSite: 'lax',
        maxAge: SESSION_COOKIE_MAX_AGE,
        domain,
      });
    }
  }

  @Post('/login')
  @ApiResponse({ type: LoginDto })
  async login(@Body() body: LoginBody, @Res({ passthrough: true }) res: Response, @Req() req: Request) {
    const { sessionId, totpSessionId } = await this.authService.login(body);

    if (totpSessionId) {
      return { success: true, totpSessionId };
    }

    await this.setSessionCookie(res, sessionId, req);

    // Include session ID in response body for Tauri desktop app
    // (cross-origin cookies don't work in WebView2 on HTTP)
    return LoginDto.parse({ success: true, sessionId }, { reportOnly: true });
  }

  @Post('/verify-totp')
  @ApiResponse({ type: LoginDto })
  async verifyTotp(@Body() body: VerifyTotpBody, @Res({ passthrough: true }) res: Response, @Req() req: Request) {
    const { sessionId } = await this.authService.verifyTotp(body);

    await this.setSessionCookie(res, sessionId, req);

    return LoginDto.parse({ success: true, sessionId }, { reportOnly: true });
  }

  @Post('/register')
  @ApiResponse({ type: RegisterDto })
  async register(@Body() body: RegisterBody, @Res({ passthrough: true }) res: Response, @Req() req: Request) {
    const { sessionId } = await this.authService.register(body);

    await this.setSessionCookie(res, sessionId, req);

    return RegisterDto.parse({ success: true }, { reportOnly: true });
  }

  @Post('/logout')
  async logout(@Res() res: Response, @Req() req: Request) {
    res.clearCookie(SESSION_COOKIE_NAME);
    const sessionId = req.cookies[SESSION_COOKIE_NAME];

    if (!sessionId) {
      return;
    }

    await this.authService.logout(sessionId);

    return res.status(204).send();
  }

  /**
   * Start Portal OIDC (PKCE) login flow.
   *
   * Desktop opens Portal in the system browser, then returns to the Tauri app via cihub://.
   * Browser-based Hub logins continue to return directly to the Hub origin that initiated the flow.
   */
  @Get('/portal/start')
  async startPortalLogin(@Req() req: Request, @Res() res: Response, @Query('redirect_url') redirectUrl?: string, @Query('desktop') desktop?: string) {
    const proto = (req.headers['x-forwarded-proto'] as string | undefined) || req.protocol || 'http';
    const host = (req.headers['x-forwarded-host'] as string | undefined) || req.get('host');

    if (!host) {
      throw new BadRequestException('Missing host header');
    }

    const hubOrigin = `${proto}://${host}`;
    const callbackUrl = new URL('/api/auth/portal/callback', hubOrigin).toString();

    const portalBaseUrl = (this.config.get('ciCloudUrl') || '').replace(/\/+$/, '');
    if (!portalBaseUrl) {
      throw new ServiceUnavailableException('CI_CLOUD_URL is not configured on this Hub.');
    }

    const state = crypto.randomUUID();
    const codeVerifier = crypto.randomBytes(32).toString('base64url');
    const codeChallenge = crypto.createHash('sha256').update(codeVerifier).digest('base64url');
    const isDesktop = desktop === '1' || desktop === 'true';

    // Persist PKCE verifier + redirect target for the callback.
    // 10 min is plenty and avoids stale entries.
    const portalState: PortalSsoState = { codeVerifier, redirectUrl: redirectUrl || null, hubOrigin, desktop: isDesktop };
    this.cache.set(`portal_sso:${state}`, JSON.stringify(portalState), 10 * 60);

    const authorizeUrl = new URL('/api/auth/oauth2/authorize', portalBaseUrl);
    authorizeUrl.searchParams.set('client_id', 'ci-hub');
    authorizeUrl.searchParams.set('redirect_uri', callbackUrl);
    authorizeUrl.searchParams.set('response_type', 'code');
    authorizeUrl.searchParams.set('scope', 'openid email profile');
    authorizeUrl.searchParams.set('code_challenge', codeChallenge);
    authorizeUrl.searchParams.set('code_challenge_method', 'S256');
    authorizeUrl.searchParams.set('state', state);

    return res.redirect(authorizeUrl.toString());
  }

  @Get('/portal/callback')
  async portalCallback(@Req() req: Request, @Res({ passthrough: true }) res: Response, @Query('code') code?: string, @Query('state') state?: string) {
    if (!code || !state) {
      throw new BadRequestException('Missing code or state');
    }

    const portalBaseUrl = (this.config.get('ciCloudUrl') || '').replace(/\/+$/, '');
    if (!portalBaseUrl) {
      throw new ServiceUnavailableException('CI_CLOUD_URL is not configured on this Hub.');
    }

    const cached = this.cache.get(`portal_sso:${state}`);
    this.cache.del(`portal_sso:${state}`);

    if (!cached) {
      throw new BadRequestException('Invalid or expired Portal SSO state');
    }

    let codeVerifier: string;
    let redirectUrl: string | null;
    let hubOrigin: string;
    let desktop = false;

    try {
      const parsed = JSON.parse(cached) as PortalSsoState;
      codeVerifier = parsed.codeVerifier;
      redirectUrl = parsed.redirectUrl;
      hubOrigin = parsed.hubOrigin;
      desktop = parsed.desktop;
    } catch {
      throw new BadRequestException('Malformed Portal SSO state');
    }

    const callbackUrl = new URL('/api/auth/portal/callback', hubOrigin).toString();

    // Exchange code -> tokens
    const tokenUrl = new URL('/api/auth/oauth2/token', portalBaseUrl).toString();
    const body = new URLSearchParams();
    body.set('grant_type', 'authorization_code');
    body.set('client_id', 'ci-hub');
    body.set('redirect_uri', callbackUrl);
    body.set('code', code);
    body.set('code_verifier', codeVerifier);

    const tokenRes = await axios.post(tokenUrl, body.toString(), {
      headers: { 'content-type': 'application/x-www-form-urlencoded' },
      // better-auth sometimes uses cookies; but token exchange should be pure.
      validateStatus: () => true,
    });

    if (tokenRes.status < 200 || tokenRes.status >= 300) {
      this.logger.warn('Portal token exchange failed', { status: tokenRes.status, data: tokenRes.data });
      throw new UnauthorizedException('Portal token exchange failed');
    }

    const accessToken = (tokenRes.data as { access_token?: string } | undefined)?.access_token;
    if (!accessToken) {
      this.logger.warn('Portal token exchange missing access_token', { data: tokenRes.data });
      throw new UnauthorizedException('Portal token exchange missing access token');
    }

    // Fetch userinfo
    const userinfoUrl = new URL('/api/auth/oauth2/userinfo', portalBaseUrl).toString();
    const userinfoRes = await axios.get(userinfoUrl, {
      headers: { authorization: `Bearer ${accessToken}` },
      validateStatus: () => true,
    });

    if (userinfoRes.status < 200 || userinfoRes.status >= 300) {
      this.logger.warn('Portal userinfo fetch failed', { status: userinfoRes.status, data: userinfoRes.data });
      throw new UnauthorizedException('Portal userinfo request failed');
    }

    const email = (userinfoRes.data as { email?: string } | undefined)?.email;
    if (!email) {
      this.logger.warn('Portal userinfo missing email', { data: userinfoRes.data });
      throw new UnauthorizedException('Portal userinfo missing email');
    }

    const operator = await this.userRepository.getFirstOperator();
    if (!operator) {
      throw new ForbiddenException('No operator user is configured on this Hub');
    }

    if (operator.username.trim().toLowerCase() !== email.trim().toLowerCase()) {
      // For now, only allow Portal login for the operator email already configured on this Hub.
      // This avoids silently elevating a Portal user to local admin.
      throw new UnauthorizedException('Portal account does not match this Hub operator');
    }

    const sessionId = await this.sessionManager.createSession(operator.id);
    await this.setSessionCookie(res, sessionId, req);

    if (desktop) {
      const desktopToken = crypto.randomUUID();
      const exchangePayload: PortalDesktopExchange = {
        sessionId,
        redirectPath: toDesktopRedirectPath(redirectUrl, hubOrigin),
      };
      this.cache.set(`portal_sso_desktop:${desktopToken}`, JSON.stringify(exchangePayload), 60);
      return res.redirect(buildPortalDesktopDeepLink(desktopToken));
    }

    // Redirect back to the requested URL if it's same-origin; otherwise go home.
    const safeRedirect = resolveSameOriginRedirectUrl(redirectUrl, hubOrigin);
    if (safeRedirect) {
      return res.redirect(safeRedirect);
    }

    return res.redirect('/home');
  }

  @Get('/portal/desktop-exchange')
  @ApiResponse({ type: PortalDesktopExchangeDto })
  async exchangePortalDesktopLogin(@Query('token') token?: string) {
    if (!token) {
      throw new BadRequestException('Missing desktop exchange token');
    }

    const cacheKey = `portal_sso_desktop:${token}`;
    const cached = this.cache.get(cacheKey);
    this.cache.del(cacheKey);

    if (!cached) {
      throw new BadRequestException('Invalid or expired desktop exchange token');
    }

    try {
      const parsed = JSON.parse(cached) as PortalDesktopExchange;
      return PortalDesktopExchangeDto.parse(parsed, { reportOnly: true });
    } catch {
      throw new BadRequestException('Malformed desktop exchange payload');
    }
  }

  @Patch('/username')
  @UseGuards(AuthGuard)
  async changeUsername(@Body() body: ChangeUsernameBody, @Req() req: Request, @Res() res: Response) {
    const userId = req.user?.id;

    if (!userId) {
      throw new TranslatableError('SYSTEM_ERROR_YOU_MUST_BE_LOGGED_IN');
    }

    await this.authService.changeUsername({ userId, ...body });

    res.clearCookie(SESSION_COOKIE_NAME);
    return res.status(204).send();
  }

  @Patch('/password')
  @UseGuards(AuthGuard)
  async changePassword(@Body() body: ChangePasswordBody, @Req() req: Request, @Res() res: Response) {
    const userId = req.user?.id;

    if (!userId) {
      throw new TranslatableError('SYSTEM_ERROR_YOU_MUST_BE_LOGGED_IN');
    }

    await this.authService.changePassword({ userId, ...body });

    res.clearCookie(SESSION_COOKIE_NAME);
    return res.status(204).send();
  }

  @Patch('/totp/get-uri')
  @UseGuards(AuthGuard)
  @ApiResponse({ type: GetTotpUriDto })
  async getTotpUri(@Body() body: GetTotpUriBody, @Req() req: Request) {
    const userId = req.user?.id;

    if (!userId) {
      throw new TranslatableError('SYSTEM_ERROR_YOU_MUST_BE_LOGGED_IN');
    }

    const res = await this.authService.getTotpUri({ userId, ...body });
    return GetTotpUriDto.parse(res, { reportOnly: true });
  }

  @Patch('/totp/setup')
  @UseGuards(AuthGuard)
  async setupTotp(@Body() body: SetupTotpBody, @Req() req: Request) {
    const userId = req.user?.id;

    if (!userId) {
      throw new TranslatableError('SYSTEM_ERROR_YOU_MUST_BE_LOGGED_IN');
    }

    await this.authService.setupTotp({ userId, totpCode: body.code });
  }

  @Patch('/totp/disable')
  @UseGuards(AuthGuard)
  async disableTotp(@Body() body: DisableTotpBody, @Req() req: Request) {
    const userId = req.user?.id;

    if (!userId) {
      throw new TranslatableError('SYSTEM_ERROR_YOU_MUST_BE_LOGGED_IN');
    }

    await this.authService.disableTotp({ userId, ...body });
  }

  @Post('/reset-password')
  @ApiResponse({ type: ResetPasswordDto })
  async resetPassword(@Body() body: ResetPasswordBody) {
    const { email } = await this.authService.changeOperatorPassword(body);

    return ResetPasswordDto.parse({ success: true, email }, { reportOnly: true });
  }

  @Delete('/reset-password')
  async cancelResetPassword() {
    await this.authService.cancelPasswordChangeRequest();
  }

  @Get('/reset-password')
  @ApiResponse({ type: CheckResetPasswordRequestDto })
  async checkResetPasswordRequest() {
    const isPending = await this.authService.checkPasswordChangeRequest();

    return CheckResetPasswordRequestDto.parse({ isRequestPending: isPending }, { reportOnly: true });
  }

  @Post('/password-reset/request')
  @ApiResponse({ type: PasswordResetRequestDto })
  async requestPasswordReset(@Body() body: PasswordResetRequestBody, @Req() req: Request) {
    await this.authService.requestPasswordReset({ email: body.email, ipAddress: req.ip });

    return PasswordResetRequestDto.parse(
      { success: true, message: 'If this email is registered, you will receive reset instructions shortly.' },
      { reportOnly: true },
    );
  }

  @Get('/password-reset/verify/:token')
  @ApiResponse({ type: PasswordResetVerifyResponseDto })
  async verifyPasswordResetToken(@Req() req: Request) {
    const token = String(req.params.token ?? '');
    const result = await this.authService.verifyPasswordResetToken(token);
    return PasswordResetVerifyResponseDto.parse(result, { reportOnly: true });
  }

  @Post('/password-reset/complete')
  @ApiResponse({ type: PasswordResetCompleteDto })
  async completePasswordReset(@Body() body: PasswordResetCompleteBody, @Req() req: Request) {
    await this.authService.completePasswordReset({ token: body.token, newPassword: body.newPassword, ipAddress: req.ip });

    return PasswordResetCompleteDto.parse(
      { success: true, message: 'Password updated. You can now log in with your new password.' },
      { reportOnly: true },
    );
  }

  @Get('/traefik')
  async traefik(@Req() req: Request, @Res() res: Response) {
    if (req.user) {
      this.logger.debug('User authenticated for Traefik forward auth', { username: req.user.username });
      res.setHeader('X-CI-Hub-User', req.user.username);
      return res.status(200).send();
    }

    const uri = req.headers['x-forwarded-uri'] as string;
    const proto = req.headers['x-forwarded-proto'] as string;
    const host = req.headers['x-forwarded-host'] as string;

    this.logger.debug('Unauthenticated Traefik forward auth request', { uri, proto, host });

    const subdomains = host.split('.');
    const app = subdomains[0] ?? '';
    const rootDomain = subdomains.slice(1).join('.');

    const redirectUrl = new URL(uri, `${proto}://${host}`);

    const loginUrl = new URL('/login', `${proto}://${rootDomain}`);
    loginUrl.searchParams.set('redirect_url', redirectUrl.toString());
    loginUrl.searchParams.set('app', app);

    this.logger.debug('Redirecting to login', { loginUrl: loginUrl.toString(), redirectUrl: redirectUrl.toString(), app });

    return res.status(302).redirect(loginUrl.toString());
  }
}